+++ source/service/loader/assemblers/operations.js
@@ -14,7 +14,8 @@ let fileRe = /^file:\//;
 
 export default function(config)
 {
-    let { assetTypes, root, tagLegend, toAssetId, toAssetName, toExampleFile, useFilenameOnly } = config;    
+    let { assetTypes, root, tagLegend, toAssetId, toAssetName, toExampleFile, useFilenameOnly } = config;  
+
     let types = Object.entries(assetTypes);
     let tagmap = new Map();
 
@@ -22,6 +23,7 @@ export default function(config)
 
 
     let op = {};
+
 
     /**
         Processes assets for `record`.
@@ -36,22 +38,33 @@ export default function(config)
     {
         let { name } = record;
 
-        return async ({ assemble }) =>
+        return async ({ assemble, assets }) =>
         {
-            let mapTypes = async ([ tid, { plural, singular } ]) => 
+            record.assets = [];
+
+            let mapType = async ([ tid, { plural, singular } ]) => 
             {
-                let files = await globit(record[plural], root);
-                let mapFiles = path => assemble[singular]({ path, section: name, tid })
-                let filterFiles = list => list.filter(is).map(item => item.uid)
+                let mapFile = async path =>
+                {
+                    let asset = await assemble[singular]({ path, section: name, tid });
+
+                    if (asset)
+                    {
+                        let { uid } = asset;
+
+                        if (assets[uid])
+                            log.warn(`duplicate asset {:emph:${uid}} was skipped`);
+                        else
+                            (record.assets.push(uid), assets[uid] = asset);
+                    }
+                }
+
+                await globit(record[plural], root).then(files => Promise.all(files.map(mapFile)));
 
                 delete record[plural];
-
-                return Promise.all(files.map(mapFiles)).then(filterFiles);
             }
 
-            record.assets = await Promise.all(types.map(mapTypes)).then(array => array.flat());
-
-            return record;
+            return Promise.all(types.map(mapType)).then(() => record);
         }
     }    
 
@@ -72,6 +85,7 @@ export default function(config)
 
         return record;
     }
+
 
     /**
         Parse `example` markdown content into `record`.
@@ -162,10 +176,10 @@ export default function(config)
 
         if (content)
         {
-            return ({ files, td }) => 
+            return ({ files, md }) => 
             {
                 let mcid = pascalCase(`Article_${uid}`);
-                let { doc, matter } = td.content.parse(content, { vars: { uid } });
+                let { doc, matter } = md.content(content, { vars: { uid } });
 
                 files[np.join(paths.components, 'articles', `${mcid}.svt`)] = doc;
 
@@ -310,13 +324,13 @@ export default function(config)
 
         if (is.nonao(path))
         {            
-            return async ({ parsers, td }) =>
+            return async ({ parsers, md }) =>
             {
                 let parser = { ...parsers['*'], ...parsers[path.ext] };
 
                 if (parser.use)
                 {
-                    let doxer = doxie[tid](path.path, td.comment);
+                    let doxer = doxie[tid](path.path, md.comment);
                     // doxer validates all data from parser
                     doxer.asset = await parser.use(path.abs, docson);
 
@@ -419,37 +433,6 @@ export default function(config)
         return record;
     }
 
-
-    /**
-        Tracks record with UID in master list.
-
-        @param { object } record
-          - `uid`: identifier for `record`
-        @return { object }
-          Asset record or `null` if asset already added.
-    */
-    op.track = async (record, exec) =>
-    {
-        record = await exec.identify(record);
-
-        let { uid } = record;
-
-        if (uid)
-        {
-            return ({ assets }) =>
-            {
-                if (assets[uid])
-                {
-                    log.warn(`duplicate asset {:emph:${uid}} was skipped`);
-                    return null;
-                }
-
-                return assets[uid] = record;
-            }
-        }
-
-        return record;
-    }
 
     return op;
 }