+++ source/utils/confine/index.js
@@ -11,11 +11,6 @@ import helpers from './helpers.js'
     It does not provide control with array mutator operations (push, 
     splice, unshift, etc.).
 
-    `assign` function will receive
-    - `apply` - pass this function the value to be assigned
-    - `value` - the original value to be assigned
-    - `at` - object with spec `path` and actual `name` for value assignment 
-
     If `apply` throws, no value was assigned.
 
     @param { object } prints
@@ -25,7 +20,8 @@ import helpers from './helpers.js'
 */
 export default function (prints, assign)
 {
-    assign ||= (apply, value) => apply(value)
+    assign ||= ({ apply, spec, value, verify }) => apply(verify(spec, value))
+    // assign ||= (apply, value) => apply(value)
 
     /**
         Governs object manipulation to adhere to a set of (blue) `prints`.
@@ -76,16 +72,14 @@ export default function (prints, assign)
             return prints[at.path] ? toSpec(at) : toSpec(id('*', prop, value));
         }
 
-        let useDefault = spec => is.undef(spec.default) ? void 0 : verify(spec, spec.default)
-
         let verifyItems = (target, values, start = 0) => 
         {
             let reducer = (array, value, index) =>
             {
                 let spec = propSpec(target, start + index, value);
-                let apply = value => array.push(verify(spec, value));
+                let apply = value => array.push(value);
 
-                assign(apply, value, spec.at);
+                assign({ apply, spec, value, verify });
                 return array;
             }
 
@@ -121,8 +115,12 @@ export default function (prints, assign)
 
                 if (is.undef(target[prop]))
                 {
-                    let defval = useDefault(propSpec(target, prop));
-                    if (!is.undef(defval)) target[prop] = defval;
+                    let spec = propSpec(target, prop);
+
+                    if (spec.immutable === true) 
+                        target[prop] = spec.default;
+                    else if (!is.undef(spec.default)) 
+                        target[prop] = verify(spec, spec.default);
                 }
 
                 return target[prop];
@@ -131,19 +129,25 @@ export default function (prints, assign)
             set(target, prop, value)
             {
                 let spec = propSpec(target, prop, value);
-                let apply = value => target[prop] = verify(spec, value);
+                let apply = value => target[prop] = value;
 
-                assign(apply, value, spec.at);
+                assign({ apply, spec, value, verify });
                 return true;
             }
         });
     }
 
-    let verify = ({ at, merge, prop, target, test }, value) =>
+    let verify = ({ at, immutable, merge, prop, target, test }, value) =>
     {
-        let result = is.func(test) ? test(helpers(at.name, value)) : null;
+        let result = null;
+
+        if (immutable === true)
+            result = new AcidValidateError('value cannot be changed', at.name);
+        else if (is.func(test))
+            result = test(helpers(at.name, value));
+
         // throw if validation error message returned
-        if (is.string(result)) throw new AcidValidateError(result);
+        if (result instanceof AcidValidateError) throw result;
 
         let final = value = is.func(result) ? result() : value;