+++ source/utils/of.js
@@ -41,6 +41,9 @@ export default function (value)
 
         // objects
         case 'object':
-            return value === null ? 'null' : Array.isArray(value) ? 'array' : 'nonao'
+            if (value === null) return 'null';
+            if (Array.isArray(value)) return 'array';
+            if (value.constructor === Object || !value.constructor) return 'plain';
+            return 'nonao';
     }
 }