+++ changelog.md
@@ -7,6 +7,23 @@
 
 
 ---
+## v0.0.4
+
+- add fenced block isolation page
+- add `updateMarkdown` option to manipulate markdown content before parsing
+- add `finalizeAsset` option to update assets before build serialization
+- add option to hide watermark logo in docsite
+- update documentation for internal components
+- add rank for tag ordering in `tagLegend` option
+- add `tinfo` client utility for tag details
+- filter component props with `@ignore` tags from the build
+- add fenced block `file` option to load file content
+- allow plugin functions to be async
+- update `socket` and `server` config options to automatically find open port
+- add immutable properties for `confine` utility
+- change `root` option to be immutable
+
+
 ## v0.0.3
 
 - renderer extension functions no longer receive `modulize` and `partition` props